Extract styles
